requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
contourpy==1.0.5
cycler==0.11.0
Django==3.0.4
django-annoying==0.10.6
django-auth-ldap==2.1.0
django-filter==2.2.0
django-jsonstore==0.4.1
django-viewflow==1.6.0
fonttools==4.37.2
GDAL @ file:///C:/Users/sserieux/Downloads/GDAL-3.4.3-cp38-cp38-win32.whl
httplib2==0.17.0
idna==2.8
img2pdf==0.3.3
kiwisolver==1.4.4
lxml==4.4.2
matplotlib==3.6.0
numpy==1.23.1
ocrmypdf==9.5.0
overpy==0.6
packaging==21.3
pandas==1.4.3
pdf2image==1.11.0
pdfminer.six==20200104
pikepdf==1.9.0
Pillow==7.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.4
pyodbc==4.0.34
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pytesseract==0.3.1
python-dateutil==2.8.1
python-ldap==3.2.0
pytz==2022.1
reportlab==3.5.34
requests==2.22.0
six==1.14.0
sortedcontainers==2.1.0
sqlparse==0.3.1
tk==0.1.0
tqdm==4.41.1
urllib3==1.25.8
XlsxWriter==3.0.3
